requirements.txt
fastapi==0.82.0
uvicorn==0.18.2
httpx==0.23.0
pydub==0.25.1
SpeechRecognition==3.8.1
python-dotenv==0.21.0
pillow
requests==2.28.1
Jinja2==3.1.2
watchdog==2.2.1
python-multipart
openai==0.28.0